+++ src/lib/project/createProject.ts
@@ -5,6 +5,8 @@ import stripInvalidFilenameCharacters from "shared/lib/helpers/stripInvalidFilen
 import { ERR_PROJECT_EXISTS, projectTemplatesRoot } from "consts";
 import copy from "lib/helpers/fsCopy";
 import { getGlobalPluginsPath } from "lib/pluginManager/globalPlugins";
+import { Value } from "@sinclair/typebox/value";
+import { ProjectMetadataResource } from "shared/lib/resources/types";
 
 export interface CreateProjectInput {
   name: string;
@@ -32,24 +34,29 @@ const createProject = async (options: CreateProjectInput) => {
     : path.join(projectTemplatesRoot, options.template);
   const projectTmpDataPath = `${projectPath}/project.gbsproj`;
   const projectDataPath = `${projectPath}/${projectFolderName}.gbsproj`;
+  const templateDataPath = path.join(templatePath, "project.gbsproj");
+
   const { username } = os.userInfo();
 
   if (fs.existsSync(projectPath)) {
     throw ERR_PROJECT_EXISTS;
   }
 
+  const projectMetadata = (await fs.readJSON(templateDataPath)) as unknown;
+  if (!Value.Check(ProjectMetadataResource, projectMetadata)) {
+    throw new Error("Template project.gbsproj is invalid");
+  }
+
+  projectMetadata.name = projectFolderName;
+  projectMetadata.author = username;
+
   await fs.ensureDir(projectPath);
   await copy(templatePath, projectPath, {
     ignore: shouldIgnoreTemplatePath(templatePath),
   });
 
-  // Replace placeholders in data file
-  const dataFile = (await fs.readFile(projectTmpDataPath, "utf8"))
-    .replace(/___PROJECT_NAME___/g, projectFolderName)
-    .replace(/___AUTHOR___/g, username);
-
   await fs.unlink(projectTmpDataPath);
-  await fs.writeFile(projectDataPath, dataFile);
+  await fs.writeJSON(projectDataPath, projectMetadata, { spaces: 2 });
   return projectDataPath;
 };